refactor(step4): 清理废弃代码 — 删除 moduleTreeDataProvider 和 csmlogHoverTranslations
删除：
- src/moduleManager/moduleTreeDataProvider.ts — 已废弃的 TreeDataProvider
  （生产代码仅使用其 ViewState 和 ModuleTreeItem 类型）
- src/csmlogHoverTranslations.ts — Step 3 中已迁移到 hoverData/translations.ts

新增：
- src/moduleManager/moduleTreeTypes.ts — 从废弃文件中提取 ViewState 和
  ModuleTreeItem 类型，供 sidebar view 和 controller 继续使用

所有单元测试通过。
